PCB splitted
In order to use the debug support, it has been needed to add the ISP port. But it means to loose 4 pins.
So, I have added a I/O extender chip on I2C bus (other 2 pins lost). Because the complexity has encreased
I have had to split the PCB in thwo different ones.