Fix init_MCP23008()
The previous version functioned very well with firmware6-test but not for firmware5-test.
In fact, this last failed at the initialization step. But the first part of the two files
are almost the same!!!! The problem was about the IOCON-register setting. In fact, after
its configuration the sequential access "should be" disabled, and reselecting the same
regiter should be not required. In fact, it function very well with firmware6-test file,
but NOT with firmware5-test!!! For this reason I added a IOCON selecting before the reading
op. In this way everything function!!!!!