Unrecognized char trouble fixed
The integration tests have shown a problem on the first received char. This invisible one, make  (often but not
always) fail the pin definitioon syntax matching. I was unable to discover this character, in Perl, (Perl is not C), so
I modified the regex to be a more generic one. It is just a workaround. But, at the moment, it functions very well