i_DOWNLIGHT removed
The traditional old type left control has a selector to control the down/up light. In fact, starting to develop the code
using this device type, I defined i_DOWNLIGHT input pin to control the down-light status. But, because I broke the device
I searched for a better solution, and I find a new device type where the switch has been replaced with a button for the
up-light only. Because I don't manage the day-light, this solution bring me just advantages: I can save a pin, and I can
provide a better using for the light-on/off using