Rollback
It will be the last version using DIP. I have to accept it is an old technology, this PCB would be a great result in the 2000 age
but a century quarter is past! :-( It is the SMD age, now. So, after the second upograde of my laboratory,  I will rebuild the
PCB using SMDs.