Pin-header connectors adjustment
The main idea to mount the two pcb-A on the pcbB has been dropped, because the result is a too much chatic one and I realized that
wires welder on the top is a dangerous solution, they can create unwanted contacts and short-circuits... So, I choise to place the
two PCB beside. So, in order to get a more ordered and clean solution, I am using flat cables instead wires, and pin-header
connectors instead of welding them directly. I foung good connectors where every female-plug is blocked in its male-socket. But,
they are available just for even numbers of pins. For this reason, I have adjusted all the pin-header connectors and their
positions.