The C version has replaced Perl one
The Perl version started to have problems: it read corrupted chars from the serial console. I have used that version for long time
without problems and I don't know why it started to get corrupted chars. However, rewriting it in C was one of the items in my todo
list, so I take the opportunity to do it. In fact Perl version has been always too slow, and the screen was no stable.
The C version is really better.

A special huge thanks to Brian Kernighan & Dennis Ritchie
(You are one of the stars that drive my path, always)