i_LIGHTONOFF pin swapped with i_STARTBUTTON one
The pin-0 (where is connected i_LIGHTONOFF) is used by the MCU to select which boot it has to perform. It means, if you turn-on
your motorbike with the light turned-on, then the motorbike will be frozen. I have swapped with i_STARTBUTTON, because nobody
turned-on his netal-horse with the start button already pushed!!!!