New version based on NCurses
The main problem of this console has always been the screen flickering; because the display is rewritten every time the
console read a new log, the displayed text is not stable and does not allow the user to read it in easyly way. For this
reason I rewrote this software from Perl to C. The C version is absolutely better, but not yet enough.

So, in order to achieve a better result, I have used NCurses library. The result is quiet better, but not enough, So, in
the future I will convert the printing method from a sequential way to a positioonal one