Size changes and cover added
In order to close the cover I have planed to use metal clothespins because they are very common on motorbikes and they ensure
a high level of grip. But for the size I used (5mm skrew) they can enter in the internal box space of 10mm. Considered the
(25mm) PCB high the box size become inadequate. So, this is the reason because the new box internal hight is 40mm.