Fix: cables exit hole
In order to find an usable cables exit-hole I turned the box upside down. Now, the cover will be fixed on the motorbike
framework and the bottom will take the koso cockpit.
This version is not easy to weller, but it allows me to find enough space inside. In the cover locker the two cables exit
holes have 8mm diameter instead of 16mm, it is required to avoid box deformation during the weller procedure. You have to
modify the holes after.